requirements.txt
pyTelegramBotAPI==4.12.0
Pillow==10.0.0
APScheduler==3.10.1